pyproject.toml
[project]
name = "llmp"
version = "0.1.0"
description = ""
authors = [
    {name = "MS",email = "mykola_shumskiy@outlook.pt"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.2,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
